Add files via upload
The data set explained and goals

Every project starts with data and with goals. So let's first look into the data. The data that we're going to use for our project to contact trace people is a combination of two datasets. There is a public dataset for San Francisco where you have businesses in San Francisco. The instructor used this dataset, and created fake users for this dataset. So he created 5,000 people, and  selected 2,500 stores out of the San Francisco dataset. Then let these 5,000 people make in summary 50,000 transactions, or 50,000 visits, to these 2,500 stores. The information in the CSV file of this project is in every row of this file. We have the information for the user name, in this case, as an example, Casey Sanders. We have the deviceID of Casey, which is a unique ID. And we have the scan timestamp as an epoch value when this person has visited the business. Then we have a unique business ID, the name of the business, the actual address in San Francisco, the zip code, and a longitude and latitude value. So if we would want to do this, we could get up a map of San Francisco and put the businesses onto the map. With that, we have our data. Our main goal here in this project is to set up a database where we put this data in and we can select an individual person, see where this person went, and see who this person met there, or who else was at this business. So let's start designing our database.